Add two new environment variables:
 o SIPLOG_TFORM={abs, rel}
 o SIPLOG_TSTART=float(x.y)

When SIPLOG_TFORM is set to "rel" the timestamp reported is going
to be counting from either when SipLogger class has been inited
if no SIPLOG_TSTART is defined, or from the timestamp provided by
the SIPLOG_TSTART value. This is mostly to aid post-mortem debugging
of CI tests failures, so that we can run bunch of scripts and get
them report timestamps that are easy to understand and consistent
between different components.